+++ Sources/AtprotoOAuth/Agent/AtprotoOAuthAgent.swift
@@ -10,54 +10,66 @@ import AtprotoTypes
 import Foundation
 import GermConvenience
 import HTTPTypes
-import OAuth
+import OAuth4Swift
 
 public actor AtprotoOAuthAgent {
 	public nonisolated let repo: Atproto.DID
 	public nonisolated let authenticatedDID: Atproto.DID
 	public nonisolated let resolver: Atproto.Resolver
-	public let clientMetadata: OAuthClient
+	public let clientId: String
 	public let authFetcher: HTTPFetcher
 
 	private let nonceCache: NSCache<NSString, IndexedNonce> = NSCache()
 
 	enum State {
-		case active(SessionState)
+		case active(OAuth.SessionState)
 		case expired
 
-		init(archive: SessionState.Archive?) {
+		init(archive: OAuth.SessionState.Archive?) throws {
 			if let archive {
-				self = .active(.init(archive: archive))
+				self = .active(try .init(archive: archive))
 			} else {
 				self = .expired
 			}
 		}
 	}
 	var state: State
+	//concurrency workaround to store the key held in SessionState
+	private let _dpopKey: DPoPKey?
 	public var lazyServerMetadata: LazyResource<AuthServerMetadata>
-	public var refreshTask: Task<SessionState.Mutable, Error>?
+	public var lazyIssuer: LazyResource<URL>
+	public var refreshTask: Task<OAuth.SessionState.TokenState, Error>?
 
-	private let saveStream: AsyncStream<SessionState.Mutable?>
-	private let saveContinuation: AsyncStream<SessionState.Mutable?>.Continuation
+	private let saveStream: AsyncStream<OAuth.SessionState.Archive.Mutable?>
+	private let saveContinuation: AsyncStream<OAuth.SessionState.Archive.Mutable?>.Continuation
 	public enum StateUpdate {
 		case loggedOut
 	}
 	public let updateStream: AsyncStream<StateUpdate>
 	private let updateContinuation: AsyncStream<StateUpdate>.Continuation
+	
+	private let clientAuth = OAuth.ClientAuth.None()
 
 	private init(
 		did: Atproto.DID,
-		clientMetadata: OAuthClient,
+		clientId: String,
 		state: State,
 		authFetcher: HTTPFetcher,
 		atprotoResolver: Atproto.Resolver
 	) {
 		self.repo = did
 		self.authenticatedDID = did
-		self.clientMetadata = clientMetadata
+		self.clientId = clientId
 		self.state = state
 		self.authFetcher = authFetcher
 		self.resolver = atprotoResolver
+		
+		switch state {
+		case .active(let sessionState):
+			_dpopKey = sessionState.dPopKey
+		case .expired:
+			_dpopKey = nil
+		}
 
 		self.lazyServerMetadata = .init(
 			fetchTaskGenerator: {
@@ -69,13 +81,13 @@ public actor AtprotoOAuthAgent {
 						)
 					return try await authFetcher.authServerDiscovery(
 						issuer: authorizationServerURL
-					).tryUnwrap
+					)
 				}
 			})
 
 		nonceCache.countLimit = 25
 
-		(saveStream, saveContinuation) = AsyncStream<SessionState.Mutable?>
+		(saveStream, saveContinuation) = AsyncStream<OAuth.SessionState.Archive.Mutable?>
 			.makeStream(bufferingPolicy: .bufferingNewest(1))
 
 		(updateStream, updateContinuation) = AsyncStream<StateUpdate>
@@ -85,10 +97,15 @@ public actor AtprotoOAuthAgent {
 	//propagate new state to our in-memory opject properties
 	//then through the async streams
 
-	private func save(sessionMutable: SessionState.Mutable) throws {
-		try session.updated(mutable: sessionMutable)
+	private func save(tokenState: OAuth.SessionState.TokenState) throws {
+		try session.updated(tokenState: tokenState)
 
-		saveContinuation.yield(sessionMutable)
+		saveContinuation.yield(
+			.init(
+				clientAuth: try session.authArchive,
+				tokenState: tokenState
+			)
+		)
 		//don't need to undestand refresh in the UI yet
 		//		updateContinuation.yield( )
 	}
@@ -104,9 +121,9 @@ public actor AtprotoOAuthAgent {
 extension AtprotoOAuthAgent {
 	public struct Archive: Sendable, Codable {
 		let did: String
-		public let session: SessionState.Archive?
+		public let session: OAuth.SessionState.Archive?
 
-		public init(did: String, session: SessionState.Archive?) {
+		public init(did: String, session: OAuth.SessionState.Archive?) {
 			self.did = did
 			self.session = session
 		}
@@ -114,13 +131,16 @@ extension AtprotoOAuthAgent {
 
 	static func restore(
 		archive: Archive,
-		clientMetadata: OAuthClient,
+		clientId: String,
 		authFetcher: HTTPFetcher,
 		atprotoResolver: Atproto.Resolver
-	) throws -> (AtprotoOAuthAgent, AsyncStream<SessionState.Mutable?>) {
+	) throws -> (
+		AtprotoOAuthAgent,
+		AsyncStream<OAuth.SessionState.Archive.Mutable?>
+	) {
 		let session = try AtprotoOAuthAgent(
 			archive: archive,
-			clientMetadata: clientMetadata,
+			clientId: clientId,
 			authFetcher: authFetcher,
 			atprotoResolver: atprotoResolver
 		)
@@ -129,13 +149,13 @@ extension AtprotoOAuthAgent {
 
 	private init(
 		archive: Archive,
-		clientMetadata: OAuthClient,
+		clientId: String,
 		authFetcher: HTTPFetcher,
 		atprotoResolver: Atproto.Resolver
 	) throws {
 		try self.init(
 			did: .init(string: archive.did),
-			clientMetadata: clientMetadata,
+			clientId: clientId,
 			state: .init(archive: archive.session),
 			authFetcher: authFetcher,
 			atprotoResolver: atprotoResolver
@@ -143,11 +163,13 @@ extension AtprotoOAuthAgent {
 	}
 
 	//if expired not worth saving
-	var archive: SessionState.Archive? {
-		guard case .active(let sessionState) = state else {
-			return nil
+	var archive: OAuth.SessionState.Archive? {
+		get throws {
+			guard case .active(let sessionState) = state else {
+				return nil
+			}
+			return try sessionState.archive
 		}
-		return sessionState.archive
 	}
 }
 
@@ -167,8 +189,12 @@ extension AtprotoOAuthAgent: AuthPDSAgent {
 	}
 }
 
-extension AtprotoOAuthAgent: OAuthSessionCapabilities {
-	public var session: SessionState {
+extension AtprotoOAuthAgent: OAuth.SessionCapabilities {
+	public func refreshed(tokenState: OAuth.SessionState.TokenState) throws {
+		try save(tokenState: tokenState)
+	}
+
+	public var session: OAuth.SessionState {
 		get throws {
 			guard case .active(let sessionState) = state else {
 				throw OAuthSessionError.sessionInactive
@@ -181,24 +207,22 @@ extension AtprotoOAuthAgent: OAuthSessionCapabilities {
 		try save(sessionMutable: sessionMutable)
 	}
 
-	public var authServerRequestOptions: AuthServerRequestOptions {
+	public var authServerRequestOptions: AuthServerRequestOptions {r
 		.atproto(
-			clientMetadata: clientMetadata,
 			did: repo,
-			authFetcher: authFetcher,
-			dpopSigner: self
+			authFetcher: authFetcher
 		)
 	}
 }
 
 extension AtprotoOAuthAgent: DPoPSigning {
-	public var dpopKey: OAuth.DPoPKey {
+	public nonisolated var dpopKey: DPoPKey {
 		get throws {
-			try session.dPopKey.tryUnwrap
+			try _dpopKey.tryUnwrap
 		}
 	}
 
-	public func getNonce(origin: String) -> OAuth.IndexedNonce? {
+	public func getNonce(origin: String) -> IndexedNonce? {
 		nonceCache.object(forKey: origin as NSString)
 	}
 
@@ -216,4 +240,23 @@ extension AtprotoOAuthAgent {
 	func getPDSUrl() async throws -> URL {
 		try await self.resolver.resolve(did: repo).pdsUrl
 	}
+}
+
+extension AtprotoOAuthAgent: OAuth.ClientAuth.Authenticable {
+	public nonisolated var tokenEndpointAuthMethod: OAuth.ClientAuth.TokenEndpointMethods {
+		.none
+	}
+
+	public func authenticate(inputs: OAuth.ClientAuth.Inputs) async throws -> (
+		FormParameters,
+		HTTPFields
+	) {
+		try await clientAuth.authenticate(clientId: clientId, inputs: inputs)
+	}
+
+	public var clientAuthArchive: Data? {
+		nil
+	}
+
+	
 }